+++ src/lambda-fns/comprehend/detect-dominant-language/index.ts
