--- src/lambda-fns/test/index.ts
@@ -1,3 +1,0 @@
-export const handler = async (event: any) => {
-  console.log(event);
-};
